+++ extensions/tuya-smart/src/utils/functions.ts
@@ -1,4 +1,4 @@
-import { DeviceCategories, DeviceCategory } from "./interfaces";
+import { Device, DeviceCategories, DeviceCategory } from "./interfaces";
 
 export const getCategory = (categories: DeviceCategory[], categoryCode: DeviceCategories): DeviceCategories => {
   const categoryInfo = categories.find((category) => category.code === categoryCode);
@@ -6,4 +6,14 @@ export const getCategory = (categories: DeviceCategory[], categoryCode: DeviceCa
     return categoryInfo.name;
   }
   return categoryCode;
+};
+
+export const isPinned = (device: Device, oldDevices: Device[]) => {
+  const oldState = oldDevices.find((oldDevice) => oldDevice.id === device.id);
+
+  if (oldState) {
+    return oldState.pinned ?? false;
+  }
+
+  return false;
 };